predicate: Remove Predicate tree related code.
This commit removes the predicate tree related code. We are making tree
an independent library. This commit is a first step in that direction.
Predicated tree can then depend on this as a completely independent
project.

Change in project includes to relative paths. This is to ensure that the
files are included correctly when added as an external dependency.