docs: Developer docs, scripts and pip requirements.
Currently the documentation does not cover the parts of developer
environment setup. This commit fixes this by adding all the notes needed
for a developer to setup the environment for doing development, testing
and documentation work.

A devsetup.sh script is added to make it easier to setup the environment
rather than manually going through the documentation.